client: fix: reubicacion de layout
